pyproject.toml
[tool.poetry]
name = "e-commerce recommendation"
version = "0.1.0"
description = ""
authors = ["Morbeus <bestnocture@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
